+++ app/exceptions/csv_errors.ts
@@ -1,0 +1,9 @@
+export class CsvImportValidationError extends Error {
+  name = 'CsvImportValidationError'
+  constructor(
+    message: string,
+    readonly errorCode: string
+  ) {
+    super(message)
+  }
+}